refactor: move View/Container splits to feature folders with PascalCase
Moves all View/Container split components from components/ (kebab-case)
to features/ (PascalCase) to match the view-container-split rule and the
existing features/chat/ pattern:

- features/nav-chats/ — NavChats pair + ConversationSidebarItem pair + 4 helpers
- features/auth/ — LoginForm pair
- features/access-request-banner/ — full folder with PascalCase renames

All import paths updated in moved files and 4 external consumers.
Biome formatting applied to features/ directory.
