+++ frontend/components/app-sidebar.tsx
@@ -1,15 +1,15 @@
-"use client";
+'use client';
 
-import type * as React from "react";
-import { Sidebar, SidebarContent } from "@/components/ui/sidebar";
-import { NavChats } from "./nav-chats";
+import type * as React from 'react';
+import { Sidebar, SidebarContent } from '@/components/ui/sidebar';
+import { NavChats } from '@/features/nav-chats/NavChats';
 
 export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
-	return (
-		<Sidebar variant="inset" {...props}>
-			<SidebarContent>
-				<NavChats />
-			</SidebarContent>
-		</Sidebar>
-	);
+  return (
+    <Sidebar variant="inset" {...props}>
+      <SidebarContent>
+        <NavChats />
+      </SidebarContent>
+    </Sidebar>
+  );
 }
