+++ frontend/features/chat/ChatView.tsx
@@ -1,55 +1,47 @@
-"use client";
+'use client';
 
-import { useEffect, useState } from "react";
-import { useStickToBottomContext } from "use-stick-to-bottom";
+import { useEffect, useState } from 'react';
+import { useStickToBottomContext } from 'use-stick-to-bottom';
+import { Message, MessageContent, MessageResponse } from '@/components/ai-elements/message';
 import {
-	Message,
-	MessageContent,
-	MessageResponse,
-} from "@/components/ai-elements/message";
+  ModelSelector,
+  ModelSelectorContent,
+  ModelSelectorEmpty,
+  ModelSelectorGroup,
+  ModelSelectorInput,
+  ModelSelectorItem,
+  ModelSelectorList,
+  ModelSelectorLogo,
+  ModelSelectorName,
+  ModelSelectorSeparator,
+  ModelSelectorTrigger,
+} from '@/components/ai-elements/model-selector';
+import { Button } from '@/components/ui/button';
+import type { AgnoMessage } from '@/lib/types';
+import { Conversation, ConversationContent } from '../../components/ai-elements/conversation';
+import { Loader } from '../../components/ai-elements/loader';
+import type { PromptInputMessage } from '../../components/ai-elements/prompt-input';
 import {
-	ModelSelector,
-	ModelSelectorContent,
-	ModelSelectorEmpty,
-	ModelSelectorGroup,
-	ModelSelectorInput,
-	ModelSelectorItem,
-	ModelSelectorList,
-	ModelSelectorLogo,
-	ModelSelectorName,
-	ModelSelectorSeparator,
-	ModelSelectorTrigger,
-} from "@/components/ai-elements/model-selector";
-import { Button } from "@/components/ui/button";
-import type { AgnoMessage } from "@/lib/types";
-import {
-	Conversation,
-	ConversationContent,
-} from "../../components/ai-elements/conversation";
-import { Loader } from "../../components/ai-elements/loader";
-import type { PromptInputMessage } from "../../components/ai-elements/prompt-input";
-import {
-	PromptInput,
-	PromptInputFooter,
-	PromptInputProvider,
-	PromptInputSubmit,
-	PromptInputTextarea,
-} from "../../components/ai-elements/prompt-input";
+  PromptInput,
+  PromptInputFooter,
+  PromptInputSubmit,
+  PromptInputTextarea,
+} from '../../components/ai-elements/prompt-input';
 
 /**
  * Props for the {@link ChatView} presentational component.
  */
 type ChatProps = {
-	/** The current message being composed by the user. */
-	message: PromptInputMessage;
-	/** Whether the assistant is generating a response (shows a loading indicator). */
-	isLoading?: boolean;
-	/** Callback fired when the textarea content changes. */
-	onUpdateMessage: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
-	/** Callback fired when the user submits a message. */
-	onSendMessage: (message: PromptInputMessage) => void;
-	/** The full conversation history to render. */
-	chatHistory: Array<AgnoMessage>;
+  /** The current message being composed by the user. */
+  message: PromptInputMessage;
+  /** Whether the assistant is generating a response (shows a loading indicator). */
+  isLoading?: boolean;
+  /** Callback fired when the textarea content changes. */
+  onUpdateMessage: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
+  /** Callback fired when the user submits a message. */
+  onSendMessage: (message: PromptInputMessage) => void;
+  /** The full conversation history to render. */
+  chatHistory: Array<AgnoMessage>;
 };
 
 /**
@@ -60,12 +52,12 @@ type ChatProps = {
  * `useStickToBottomContext`.
  */
 const ChatScrollAnchor = ({ track: _track }: { track: number }) => {
-	const { scrollToBottom } = useStickToBottomContext();
+  const { scrollToBottom } = useStickToBottomContext();
 
-	useEffect(() => {
-		scrollToBottom();
-	}, [scrollToBottom]);
-	return null;
+  useEffect(() => {
+    scrollToBottom();
+  }, [scrollToBottom]);
+  return null;
 };
 
 /**
@@ -76,102 +68,100 @@ const ChatScrollAnchor = ({ track: _track }: { track: number }) => {
  * the parent {@link ChatContainer}.
  */
 const ChatView = ({
-	message,
-	isLoading,
-	chatHistory,
-	onSendMessage,
-	onUpdateMessage,
+  message,
+  isLoading,
+  chatHistory,
+  onSendMessage,
+  onUpdateMessage,
 }: ChatProps) => {
-	// Need to fix this.
-	const [open, setOpen] = useState(false);
-	const [selectedModel, setSelectedModel] = useState("gemini-3-flash-preview");
+  // Need to fix this.
+  const [open, setOpen] = useState(false);
+  const [selectedModel, setSelectedModel] = useState('gemini-3-flash-preview');
 
-	return (
-		<div className="overflow-hidden sm:max-w-[80%] lg:max-w-[60%] xl:max-w-[50%] mx-auto">
-			<div className="h-[90vh] flex flex-col overflow-hidden">
-				<Conversation className="flex-1 overflow-y-auto" resize="smooth">
-					<ConversationContent>
-						{chatHistory.length === 0 ? (
-							<div className="text-center my-auto font-semibold mt-8">
-								<p className="text-3xl mt-4 bg-linear-to-r from-blue-600 via-purple-600 to-pink-600 bg-clip-text text-transparent animate-gradient">
-									What can we build together?
-								</p>
-							</div>
-						) : (
-							<>
-								{chatHistory.map((message, index) => (
-									<Message from={message.role} key={`${message.role}-${index}`}>
-										<MessageContent>
-											<MessageResponse>{message.content}</MessageResponse>
-										</MessageContent>
-									</Message>
-								))}
-								{isLoading && (
-									<Message from="assistant">
-										<MessageContent>
-											<div className="flex items-center gap-2">
-												<Loader />
-												Thinking...
-											</div>
-										</MessageContent>
-									</Message>
-								)}
-							</>
-						)}
-					</ConversationContent>
-					<ChatScrollAnchor track={chatHistory.length} />
-				</Conversation>
-				<PromptInput onSubmit={onSendMessage} className="px-2 pb-2">
-					<PromptInputTextarea
-						placeholder="Ask anything about your memories or search the web..."
-						className="pr-16 bg-white min-h-12.5"
-						onChange={onUpdateMessage}
-						value={message.content}
-					/>
+  return (
+    <div className="overflow-hidden sm:max-w-[80%] lg:max-w-[60%] xl:max-w-[50%] mx-auto">
+      <div className="h-[90vh] flex flex-col overflow-hidden">
+        <Conversation className="flex-1 overflow-y-auto" resize="smooth">
+          <ConversationContent>
+            {chatHistory.length === 0 ? (
+              <div className="text-center my-auto font-semibold mt-8">
+                <p className="text-3xl mt-4 bg-linear-to-r from-blue-600 via-purple-600 to-pink-600 bg-clip-text text-transparent animate-gradient">
+                  What can we build together?
+                </p>
+              </div>
+            ) : (
+              <>
+                {chatHistory.map((message, index) => (
+                  <Message from={message.role} key={`${message.role}-${index}`}>
+                    <MessageContent>
+                      <MessageResponse>{message.content}</MessageResponse>
+                    </MessageContent>
+                  </Message>
+                ))}
+                {isLoading && (
+                  <Message from="assistant">
+                    <MessageContent>
+                      <div className="flex items-center gap-2">
+                        <Loader />
+                        Thinking...
+                      </div>
+                    </MessageContent>
+                  </Message>
+                )}
+              </>
+            )}
+          </ConversationContent>
+          <ChatScrollAnchor track={chatHistory.length} />
+        </Conversation>
+        <PromptInput onSubmit={onSendMessage} className="px-2 pb-2">
+          <PromptInputTextarea
+            placeholder="Ask anything about your memories or search the web..."
+            className="pr-16 bg-white min-h-12.5"
+            onChange={onUpdateMessage}
+            value={message.content}
+          />
 
-					<PromptInputFooter>
-						<ModelSelector open={open} onOpenChange={setOpen}>
-							<ModelSelectorTrigger asChild>
-								<Button variant="outline">
-									<ModelSelectorLogo provider="google" />
-									{selectedModel}
-								</Button>
-							</ModelSelectorTrigger>
+          <PromptInputFooter>
+            <ModelSelector open={open} onOpenChange={setOpen}>
+              <ModelSelectorTrigger asChild>
+                <Button variant="outline">
+                  <ModelSelectorLogo provider="google" />
+                  {selectedModel}
+                </Button>
+              </ModelSelectorTrigger>
 
-							<ModelSelectorContent>
-								<ModelSelectorInput placeholder="Search models..." />
-								<ModelSelectorList>
-									<ModelSelectorEmpty>No models found.</ModelSelectorEmpty>
+              <ModelSelectorContent>
+                <ModelSelectorInput placeholder="Search models..." />
+                <ModelSelectorList>
+                  <ModelSelectorEmpty>No models found.</ModelSelectorEmpty>
 
-									<ModelSelectorGroup heading="Google">
-										<ModelSelectorItem
-											value="gemini-3-flash-preview"
-											onSelect={() => {
-												setSelectedModel("gemini-3-flash-preview");
-												setOpen(false);
-											}}
-										>
-											<ModelSelectorLogo provider="google" />
-											<ModelSelectorName>
-												Gemini 3 Flash Preview
-											</ModelSelectorName>
-										</ModelSelectorItem>
-									</ModelSelectorGroup>
+                  <ModelSelectorGroup heading="Google">
+                    <ModelSelectorItem
+                      value="gemini-3-flash-preview"
+                      onSelect={() => {
+                        setSelectedModel('gemini-3-flash-preview');
+                        setOpen(false);
+                      }}
+                    >
+                      <ModelSelectorLogo provider="google" />
+                      <ModelSelectorName>Gemini 3 Flash Preview</ModelSelectorName>
+                    </ModelSelectorItem>
+                  </ModelSelectorGroup>
 
-									<ModelSelectorSeparator />
-								</ModelSelectorList>
-							</ModelSelectorContent>
-						</ModelSelector>
-						<PromptInputSubmit
-							disabled={message.content.length === 0}
-							className="absolute bottom-1 right-1 cursor-pointer"
-							status={isLoading ? "streaming" : "ready"}
-						/>
-					</PromptInputFooter>
-				</PromptInput>
-			</div>
-		</div>
-	);
+                  <ModelSelectorSeparator />
+                </ModelSelectorList>
+              </ModelSelectorContent>
+            </ModelSelector>
+            <PromptInputSubmit
+              disabled={message.content.length === 0}
+              className="absolute bottom-1 right-1 cursor-pointer"
+              status={isLoading ? 'streaming' : 'ready'}
+            />
+          </PromptInputFooter>
+        </PromptInput>
+      </div>
+    </div>
+  );
 };
 
 export default ChatView;
