+++ frontend/features/nav-chats/CollapsibleGroupHeader.tsx
