+++ frontend/features/nav-chats/SectionHeader.tsx
